fix: renamed shared folder to match sahred/services
