+++ packages/component-footer/src/footer.js
@@ -2,7 +2,7 @@
 import { shape } from "prop-types";
 import React, { useEffect } from "react";
 
-import trackReactComponent from "../../../shared/functions/componentDatalayer";
+import trackReactComponent from "../../../shared/services/componentDatalayer";
 import { Social, Legal, Innovation, Contact } from "./components";
 
 const Base = () => {
