+++ shared/index.js
@@ -1,3 +1,4 @@
 export * from "./hooks";
 export * from "./utils";
-export * from "./services/googleAnalytics"
+export * from "./services/googleAnalytics";
+export * from "./services/componentDatalayer";
