+++ shared/services/componentDatalayer.js
